match the rest of the defines for the deadzone variable

+++ src/pc/controller/controller_sdl.c
@@ -45,7 +45,7 @@ static bool joy_buttons[SDL_CONTROLLER_BUTTON_MAX ] = { false };
 static u32 mouse_buttons = 0;
 static u32 last_mouse = VK_INVALID;
 static u32 last_joybutton = VK_INVALID;
-extern s16 configStickDeadzone;
+extern u32 configStickDeadzone;
 
 static inline void controller_add_binds(const u32 mask, const u32 *btns) {
     for (u32 i = 0; i < MAX_BINDS; ++i) {
